pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-torchdebug-env"
version = "1.0.0"
description = "PyTorch Training Run Debugger - An OpenEnv environment for diagnosing and fixing broken ML training runs"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "torchdebug_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = [
    "torchdebug_env",
    "torchdebug_env.server",
    "torchdebug_env.scenarios",
    "torchdebug_env.utils",
]

[tool.setuptools.package-dir]
"torchdebug_env" = "."
"torchdebug_env.server" = "server"
"torchdebug_env.scenarios" = "scenarios"
"torchdebug_env.utils" = "utils"
